% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{COPD_proteinCodingGenes}
\alias{COPD_proteinCodingGenes}
\title{COPD_proteinCodingGenes}
\format{
A vector:
\describe{
\item{\code{value}}{Hugo Gene symbol}
}
}
\source{
<https://doi.org/10.1073/pnas.2301342120>
}
\usage{
COPD_proteinCodingGenes
}
\description{
List of genes (HGNC symbols) found associated with COPD disease
}
\keyword{datasets}
