% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{KEGG_2021_Human}
\alias{KEGG_2021_Human}
\title{KEGG_2021_Human}
\format{
A data frame with two variables:
\describe{
\item{\code{GeneSymbol}}{HUGO gene symbols belonging to a pathway}
\item{\code{PathwayName}}{Pathway name}
}
}
\source{
<https://maayanlab.cloud/Enrichr/#libraries>
}
\usage{
KEGG_2021_Human
}
\description{
KEGG pathway annotation data from Enrichr Library
}
\keyword{datasets}
