% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetFunrankR_wrapper.R
\name{NetFunrankR_wrapper}
\alias{NetFunrankR_wrapper}
\title{Title This function works as a wrapper for running the whole functional analysis using network medicine based approach.}
\usage{
NetFunrankR_wrapper(
  disease_genes = NULL,
  PathwayRanker_Type = "Both",
  PE_ORA_pop = "KEGG",
  nPermute = 10,
  string_PPI_score_th = 900,
  doPlot = TRUE,
  plot_width = 10,
  plot_height = 8,
  outdir = "./NetFunrankR_results/"
)
}
\arguments{
\item{disease_genes}{Input genes (a vector) that are used functional enrichment analysis. Default is "COPD-related genes".}

\item{PathwayRanker_Type}{A string that indicates the type of enrichment to be done. Options are: "ORA" (Hypergeometric-distribution based over-representation test), and "TW" (Topology-aware network medicine-based approach), "Both" (Default: both ORA and TW so that the results can be compared).}

\item{PE_ORA_pop}{A string that indicates the set of known annotaiton of pathways or functional term. Current version has the following options: "KEGG" (default), "Reactome" and "WikiPathway".}

\item{nPermute}{A number (default: 1000) to indicate the number of permutation iteration to be conducted for hypothesis testing in Network-medicine based functional enrichment}

\item{string_PPI_score_th}{A number (default: 900) to filter StringDB PPIs only above the given value.}

\item{doPlot}{A logical value (default: TRUE) to specify whether to plot the results.}

\item{plot_width}{A number (default: 10) to specify the output plot width.}

\item{plot_height}{A number (default: 10) to specify the output plot height.}

\item{outdir}{A string to indicate the output directory to save all the results.}
}
\value{
""
}
\description{
Title This function works as a wrapper for running the whole functional analysis using network medicine based approach.
}
\examples{
#' library(NetFunrankR)
library(dplyr)
NetFunrankR_wrapper(disease_genes = COPD_proteinCodingGenes,
              doPlot = TRUE,
              nPermute = 100)
}
