% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{protein.links}
\alias{protein.links}
\title{protein.links}
\format{
A data frame with three variables:
\describe{
\item{\code{from}}{HUGO gene symbol 1}
\item{\code{to}}{HUGO gene symbol 2}
\item{\code{combined_score}}{Combined confidence score from multiple evidences}
}
}
\source{
<https://string-db.org/>
}
\usage{
protein.links
}
\description{
Protein-Protein Interaction (human 9606) dataset from StringDB database
}
\keyword{datasets}
