% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netProximityMeasure.R
\name{quantify_network_proximity_score}
\alias{quantify_network_proximity_score}
\title{Title}
\usage{
quantify_network_proximity_score(
  neighbourhood_th,
  string_PPI_score_th,
  module_a = c(),
  module_b = c(),
  nPermute
)
}
\arguments{
\item{nPermute}{}
}
\description{
Title
}
