% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Reactome_Pathways_2024}
\alias{Reactome_Pathways_2024}
\title{Reactome_Pathways_2024}
\format{
A data frame with two variables:
\describe{
\item{\code{GeneSymbol}}{HUGO gene symbols belonging to a pathway}
\item{\code{PathwayName}}{Pathway name}
}
#' @source <https://maayanlab.cloud/Enrichr/#libraries>
}
\usage{
Reactome_Pathways_2024
}
\description{
Reactome pathway annotation data from Enrichr Library
}
\keyword{datasets}
