% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WikiPathways_2024_Human}
\alias{WikiPathways_2024_Human}
\title{WikiPathways_2024_Human}
\format{
A data frame with two variables:
\describe{
\item{\code{GeneSymbol}}{HUGO gene symbols belonging to a pathway}
\item{\code{PathwayName}}{Pathway name}
}
}
\source{
<https://maayanlab.cloud/Enrichr/#libraries>
}
\usage{
WikiPathways_2024_Human
}
\description{
WikiPathway pathway annotation data from Enrichr Library
}
\keyword{datasets}
