% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netProximityMeasure.R
\name{network.proximity}
\alias{network.proximity}
\title{Title}
\usage{
network.proximity(
  net,
  module_a_genes,
  module_b_genes,
  neighbourhood_th,
  nPermute
)
}
\arguments{
\item{nPermute}{}
}
\description{
Title
}
